refactor: ajustando o mvc
